pyproject.toml
[project]
name = "python-whatsapp-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "certifi>=2025.8.3",
    "flask>=3.1.2",
    "mongo>=0.2.0",
    "pymongo[srv]>=4.15.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
